WIP: revised filter logic
Tinkering with using python's fnmatch package to filter portions of the
hierarchy via shell-style globbing.

Essentially, added a `path` and `filtered` attribute to each of the
LookML objects and use the `path` of the objects to apply the filters
against.

As part of this tinkering I also extracted the building of the hierarchy
and filtering out of the validators module into a new hierarchy module
and itched to be able to just display it so I added a `display` option
to do that.