Join message & message_details
Also removed the check for message being a string because if we get
something else, it will cause the `join` to raise an exception, so we'll
never get to perform the check. The exception raised by join will be
caught by the caller's `try` block and a helpful "unexpected API query
result format" exception will be raised including the context of the
response we attempted to parse.