Fix bug not respecting single or hybrid mode
This commit introduces a few changes. First, it moves around the log_failing_sql so it sits in logger.py, adds a separate test module for log testing, and prepends some text before the filename so it's clear what it is in the output. Second, it fixes a bug where dimension queries were referencing their explores, not dimensions, effectively breaking single mode. It fixes another bug in cli.py where the dimension-level errors were not being printed in hybrid mode.