Entirely remove fail fast arg
After review, we decided that this behavior is beneficial on fail fast and non fail fast runs because as soon as we encounter an error, we want to move to dimension-level queries, even if we didn't catch them all. That means fail_fast for chunk queries no longer needs to be controlled.